Prevent active_support/test_case.rb from being required
If AS/test_case.rb were loaded after this, it tries to define
class AS::TestCase < ::Minitest::Test
then causes "superclass mismatch for class TestCase (TypeError)"
because we've created AS::TestCase < ::Test::Unit::TestCase here.

This dirty hack is to fake Kernel#require as if AS/test_case.rb
has been already loaded.